+++ vkcs/db/resource_cluster_with_shards.go
@@ -253,7 +253,7 @@ func ResourceDatabaseClusterWithShards() *schema.Resource {
 				Type:        schema.TypeBool,
 				Optional:    true,
 				ForceNew:    false,
-				Description: "Enable cloud monitoring for the cluster. Changing this for Redis or MongoDB creates a new instance.",
+				Description: "Enable cloud monitoring for the cluster.",
 			},
 
 			"shard": {