+++ templates/networking/resources/vkcs_networking_anycastip/import.sh
@@ -1,0 +1,1 @@
+terraform import vkcs_networking_anycastip.anycastip_1 bfbed405-dd89-41d9-aa97-6e335161146d