+++ CHANGELOG.md
@@ -8,6 +8,7 @@ description: |-
 # VKCS Provider's changelog
 
 #### v0.9.4 (unreleased)
+- Disable forced re-creation when updating cloud_monitoring_enabled for the vkcs_db_instance with MongoDB datastore type
 
 #### v0.9.3
 - Add secure_key option to cdn resource